--- MinCloud/MinCloud-Bridging-Header.h
@@ -1,17 +1,0 @@
-//
-//  MinCloud-Bridging-Header.h
-//  MinCloud
-//
-//  Created by pengquanhua on 2019/5/14.
-//  Copyright © 2019 ifanr. All rights reserved.
-//
-
-#ifndef MinCloud_Bridging_Header_h
-#define MinCloud_Bridging_Header_h
-
-
-#endif /* MinCloud_Bridging_Header_h */
-
-#import "WXApi.h"
-#import "WechatAuthSDK.h"
-#import "AlipaySDK/AlipaySDK.h"
