+++ MinCloud/module.modulemap
@@ -1,0 +1,9 @@
+module MinCloud {
+    header "WXApi.h"
+    header "WXApiObject.h"
+    header "WechatAuthSDK.h"
+    header "AlipaySDK.framework/Headers/AlipaySDK.h"
+
+    export *
+}
+
